remove map wrapper
